# docs/swagger.yaml
definitions:
  customError.AppError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  tesodev-korpes_OrderService_internal_types.Address:
    properties:
      city:
        type: string
      id:
        type: string
      state:
        type: string
      zipCode:
        type: string
    type: object
  tesodev-korpes_OrderService_internal_types.CustomerResponseModel:
    properties:
      address:
        items:
          $ref: '#/definitions/tesodev-korpes_OrderService_internal_types.Address'
        type: array
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        items:
          $ref: '#/definitions/tesodev-korpes_OrderService_internal_types.Phone'
        type: array
      updated_at:
        type: string
    type: object
  tesodev-korpes_OrderService_internal_types.Phone:
    properties:
      id:
        type: string
      phoneNumber:
        type: integer
    type: object
...
  types.CreateOrderRequestModel:
    properties:
      billing_address:
        $ref: '#/definitions/tesodev-korpes_OrderService_internal_types.Address'
      customer_id:
        type: string
      discounts:
        items:
          $ref: '#/definitions/types.Discount'
        type: array
      items:
        items:
          $ref: '#/definitions/types.OrderItem'
        type: array
      shipping_address:
        $ref: '#/definitions/tesodev-korpes_OrderService_internal_types.Address'
    required:
    - billing_address
    - customer_id
    - items
    - shipping_address
    type: object
  types.Discount:
    properties:
      discountCode:
        type: string
      endDate:
        type: string
      id:
        type: string
      role:
        type: string
      startDate:
        type: string
      type:
        type: string
      value:
        type: number
    type: object
  types.FinalPriceResult:
    properties:
      discount_applied:
        type: number
      discount_type:
        type: string
      final_price:
        type: number
      original_price:
        type: number
    type: object
...
  types.OrderItem:
    properties:
      productId:
        type: string
      productName:
        type: string
      quantity:
        type: integer
      unitPrice:
        type: number
    type: object
  types.OrderResponseModel:
    properties:
      billing_address:
        $ref: '#/definitions/tesodev-korpes_OrderService_internal_types.Address'
      created_at:
        type: string
      customer_id:
        type: string
      discounts:
        items:
          $ref: '#/definitions/types.Discount'
        type: array
      id:
        type: string
      is_delete:
        type: boolean
      items:
        items:
          $ref: '#/definitions/types.OrderItem'
        type: array
      shipping_address:
        $ref: '#/definitions/tesodev-korpes_OrderService_internal_types.Address'
      status:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
    type: object
  types.OrderWithCustomerResponse:
    properties:
      billing_address:
        $ref: '#/definitions/tesodev-korpes_OrderService_internal_types.Address'
      created_at:
        type: string
      customer:
        $ref: '#/definitions/tesodev-korpes_OrderService_internal_types.CustomerResponseModel'
      customer_id:
        type: string
      discounts:
        items:
          $ref: '#/definitions/types.Discount'
        type: array
      id:
        type: string
      is_delete:
        type: boolean
      items:
        items:
          $ref: '#/definitions/types.OrderItem'
        type: array
      shipping_address:
        $ref: '#/definitions/tesodev-korpes_OrderService_internal_types.Address'
      status:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /order/{id}/non-premium-price:
    get:
      consumes:
      - application/json
      description: Calculates and returns the final price for a non-premium order
        by order ID
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Final price details for non-premium order
          schema:
            $ref: '#/definitions/types.FinalPriceResult'
        "400":
          description: Invalid or empty order ID
          schema:
            $ref: '#/definitions/customError.AppError'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/customError.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/customError.AppError'
      summary: Get non-premium order final price
      tags:
      - orders
  /order/{id}/premium-price:
    get:
      consumes:
      - application/json
      description: Calculates and returns the final price for a premium order by order
        ID
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Final price details for premium order
          schema:
            $ref: '#/definitions/types.FinalPriceResult'
        "400":
          description: Invalid or empty order ID
          schema:
            $ref: '#/definitions/customError.AppError'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/customError.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/customError.AppError'
      summary: Get premium order final price
      tags:
      - orders
  /orders:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all orders.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns list of orders with pagination
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No orders found
          schema:
            $ref: '#/definitions/customError.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/customError.AppError'
      summary: List all orders with pagination
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order with the given data
      parameters:
      - description: Order to create
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/types.CreateOrderRequestModel'
      produces:
      - application/json
      responses:
        "201":
          description: Returns created order details
          schema:
            $ref: '#/definitions/types.OrderResponseModel'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/customError.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/customError.AppError'
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Marks the order as deleted without removing it permanently.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion success message
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/customError.AppError'
        "409":
          description: Conflict error
          schema:
            $ref: '#/definitions/customError.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/customError.AppError'
      summary: Soft delete an order by ID
      tags:
      - orders
    get:
      consumes:
      - application/json
      description: Retrieve an order with customer details by its unique ID
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order details with customer info
          schema:
            $ref: '#/definitions/types.OrderWithCustomerResponse'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/customError.AppError'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/customError.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/customError.AppError'
      summary: Get order by ID
      tags:
      - orders
  /orders/{id}/cancel:
    put:
      consumes:
      - application/json
      description: Cancel an order by its ID.
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/customError.AppError'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/customError.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/customError.AppError'
      summary: Cancel an order
      tags:
      - orders
  /orders/{id}/ship:
    put:
      consumes:
      - application/json
      description: Mark an order as shipped by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/customError.AppError'
      summary: Ship an order
      tags:
      - orders
swagger: "2.0"